backend/requirements.txt
Django
djangorestframework
django-cors-headers
validators